render level pages from one js file
